+++ src/PhpSpreadsheet/Calculation/Logical/Conditional.php
@@ -5,6 +5,7 @@ namespace PhpOffice\PhpSpreadsheet\Calculation\Logical;
 use PhpOffice\PhpSpreadsheet\Calculation\ArrayEnabled;
 use PhpOffice\PhpSpreadsheet\Calculation\Exception;
 use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\Information\ErrorValue;
 use PhpOffice\PhpSpreadsheet\Calculation\Information\ExcelError;
 use PhpOffice\PhpSpreadsheet\Calculation\Information\Value;
 
@@ -47,15 +48,16 @@ class Conditional
      */
     public static function statementIf($condition = true, $returnIfTrue = 0, $returnIfFalse = false)
     {
-        if (Value::isError($condition)) {
+        $condition = ($condition === null) ? true : Functions::flattenSingleValue($condition);
+
+        if (ErrorValue::isError($condition)) {
             return $condition;
         }
 
-        $condition = ($condition === null) ? true : (bool) Functions::flattenSingleValue($condition);
         $returnIfTrue = $returnIfTrue ?? 0;
         $returnIfFalse = $returnIfFalse ?? false;
 
-        return ($condition) ? $returnIfTrue : $returnIfFalse;
+        return ((bool) $condition) ? $returnIfTrue : $returnIfFalse;
     }
 
     /**
@@ -138,7 +140,7 @@ class Conditional
 
         $errorpart = $errorpart ?? '';
 
-        return self::statementIf(Value::isError($testValue), $errorpart, $testValue);
+        return self::statementIf(ErrorValue::isError($testValue), $errorpart, $testValue);
     }
 
     /**
@@ -164,7 +166,7 @@ class Conditional
 
         $napart = $napart ?? '';
 
-        return self::statementIf(Value::isNa($testValue), $napart, $testValue);
+        return self::statementIf(ErrorValue::isNa($testValue), $napart, $testValue);
     }
 
     /**
