+++ src/PhpSpreadsheet/Calculation/LookupRef/LookupRefValidations.php
@@ -3,8 +3,8 @@
 namespace PhpOffice\PhpSpreadsheet\Calculation\LookupRef;
 
 use PhpOffice\PhpSpreadsheet\Calculation\Exception;
+use PhpOffice\PhpSpreadsheet\Calculation\Information\ErrorValue;
 use PhpOffice\PhpSpreadsheet\Calculation\Information\ExcelError;
-use PhpOffice\PhpSpreadsheet\Calculation\Information\Value;
 
 class LookupRefValidations
 {
@@ -14,7 +14,7 @@ class LookupRefValidations
     public static function validateInt($value): int
     {
         if (!is_numeric($value)) {
-            if (Value::isError($value)) {
+            if (ErrorValue::isError($value)) {
                 throw new Exception($value);
             }
